Merge bk://bk.arm.linux.org.uk into home.transmeta.com:/home/torvalds/v2.5/linux
2002/11/02 12:57:43-08:00 tytso
[PATCH] Orlov block allocator for ext3

Here's the ext3 version.

2002/11/02 12:57:39-08:00 tytso
[PATCH] Fixup Orlov block allocator for ext2

I finally had time to look at the Orlov patches, and found a memory
leak; sbi->s_debts wasn't getting freed when the filesystem was
getting unmounted, or in the error path.

This patch also makes the following cleanups/changes:

1) Use sbi->s_debts instead of sbi->debts --- all other fields in
	struct ext2_sb_info are prefixed by "s_", so this makes things
	consistent.

2) Add support for a new inode flag, EXT2_TOPDIR_FL, which tells tells
	the Orlov allocator to treat that directory as the top of
	directory hierarchies, so that new subdirectories created in
	that directory should be spread apart.  System administrators
	should set this flag on directories like /usr/src, /usr/home, etc.

3) Add a mount-time flag, -o oldalloc, which forces the use of the old
	inode (pre-Orlov) allocator.  This makes it easier to do
	comparison benchmarks, and in case people want to use the old
	algorithm.

2002/11/02 12:47:01-08:00 torvalds
Remove last vestiges of ide-cd.c "struct packet_command". It's been
empty for a while now, and nothing uses it.

2002/11/02 11:55:44-08:00 torvalds
Add a "cmd_len" parameter to the request, so that device drivers
don't have to try to figure it out for themselves.

Make ide-cd.c use it.

BKrev: 3dc43fa1GOh-uDN0slqCJjxCPXrN-Q