Merge bk://linuxusb.bkbits.net/pci_hp-2.5 into home.transmeta.com:/home/torvalds/v2.5/linux
2002/11/05 16:21:21-08:00 greg
Merge kroah.com:/home/greg/linux/BK/bleeding_edge-2.5
into kroah.com:/home/greg/linux/BK/pci_hp-2.5

2002/11/03 00:08:36-08:00 t-kouchi
[PATCH] ACPI PCI hotplug updates

These are updates of the acpiphp driver for 2.5.
  - change debug flag from 'acpiphp_debug' to 'debug' for insmod
  - whitespace cleanup
  - message cleanup

2002/11/03 00:02:17-08:00 jung-ik.lee
[PATCH] Patch: 2.5.45 PCI Fixups for PCI HotPlug

The following patch changes function scopes only but fixes kernel dump on
Hot-Add of PCI bridge cards.

2002/11/02 23:43:53-08:00 greg
[PATCH] PCI Hotplug: fix compiler warning.

2002/11/02 23:40:07-08:00 greg
[PATCH] PCI Hotplug: removed a compiler warning of a unused variable in the cpcihp_generic driver.

2002/11/02 23:14:54-08:00 scottm
[PATCH] 2.5.45 CompactPCI driver patch 4/4

This is a patch 4 of 4 of my CompactPCI hotplug core and
drivers, consisting of the generic port I/O cPCI hotplug driver.

Let me know if the kernel parameter parsing code that's #ifndef MODULE is
objectionable.  I spent quite a while today testing it, it seems reasonably
robust.  Without it, this driver would only be useable as a module, which
I've not figured out how to do with the new kernel configuration stuff.

2002/11/02 23:07:09-08:00 scottm
[PATCH] 2.5.45 CompactPCI driver patch 3/4

This is patch 3 of 4 of my CompactPCI hotplug core and
drivers, consisting of the Ziatech ZT5550 hotplug driver.

The hardware banging code in this driver started its life in the PICMG
2.12 driver code that MontaVista released at the end of 2001.

2002/11/02 23:00:20-08:00 scottm
[PATCH] 2.5.45 CompactPCI driver patch 2/4

This is patch 2 of 4 of my CompactPCI hotplug core and
drivers, consisting of the CompactPCI hotplug driver core.

It is basically a glue layer on top of the PCI hotplug core that exposes
an API roughly similiar in concept to the API implemented by MontaVista
from the PICMG 2.12 specification, minus all the Win32isms and cruft.

2002/11/02 22:57:45-08:00 greg
PCI: move EXPORT_SYMBOL for the pbus functions to the setup-bus.c file.

This fixes a linking error if setup-bus.c isn't compiled into the kernel.

2002/11/02 22:36:38-08:00 scottm
[PATCH] 2.5.45 CompactPCI driver patch 1/4

This is a patch 1 of 4 of my CompactPCI hotplug core and
drivers, consisting of the required core PCI changes.

The various arch file changes are to change pcibios_fixup_pbus_ranges to
from __init to __devinit, so that pci_setup_bridge can be safely exported
from drivers/pci/setup-bus.c.

2002/11/02 22:25:32-08:00 rmk
[PATCH] PCI hotplug comment fixes

Fix comments about /sbin/hotplug; pci_insert_device does not call
/sbin/hotplug.

BKrev: 3dc8921dgAMZxUrVwtw3ucYlq-mUBA