[PATCH] Update macserial driver
This updates the macserial driver in 2.5 so it compiles and works.  The
main changes are to use schedule_work instead of task queues and BHs.
The patch also removes the wait_key method.

I know we need to change macserial to use the new serial
infrastructure.  I'm posting this patch in case it is useful to anyone
trying to compile up a kernel for a powermac at the moment.

BKrev: 3dc9f526x8P5T194Z9SW_5YauM2fug