[PATCH] Add nfs_writepages & backing_dev...
The following patch adds a simple ->writepages method that interprets
the extra information passed down in Andrew's writeback_control
structure, and translates it into nfs-speak.

It also adds a backing_dev_info structure that scales the readahead in
terms of the rsize. Maximum readahead is still 128k if you use 32k
rsize, but it is scaled down to 4k if you use 1k rsize.

BKrev: 3dcc30ffM2OrCzd5SdtcgSIOD2J-LA