Merge
2002/11/16 19:46:55-08:00 willy
[PATCH] eliminate pci_dev name

This removes the pci_dev "->name[]" field, and makes users use the
"struct device" name field instead.

It also trimps it down to 50 bytes and limited the length of the vendor
& device descriptions to 20 bytes each.  Now we have three cases:

 - unknown vendor & unknown device -> "PCI device %04x:%04x".
	Clearly limited to 21 chars.
 - known vendor, unknown device -> "PCI device %04x:%04x (%.20s)"
	24 + 20 chars
 - known vendor, known device -> "%.20s %.20s"
	20 + 1 + 20 + 1, EXCEPT:
 - multiple devices of the same type add " (#%d)"
	so 42 + 4 + 4 = 50.  This is the point where an IBMer tells me they
	intend to sell a machine with > 9999 PCI devices of the same type ;-)

This limits the PCI IDs to a reasonable amount of space.

For release, I think we do want to go back up to 80 for prettier output

BKrev: 3dd71180kn2nDHwOrrXyBU1E0cz7Xg