[PATCH] Forced module unload
This is the logical counterpoint to the code which marks modules
"[unsafe]" when obsolete (racy) interfaces are used.  Allows "just
remove the damn thing" rmmod -f, and taints the kernel.

Mark it dangerous and experimental in the config file to make this
doubly clear.

BKrev: 3dd7ec5eHlVIwswHUVz8lMlJf_13Kw