[PATCH] module device table restoration
Patch from Adam Richter.  I have a nicer solution based on aliases, but
it requires coordination with USB, PCI and PCMCIA maintainers, which is
taking time.

This restores the old code in the meantime: one week without this is too
long for people who need it.

BKrev: 3ddbc6e6joI_nWWRjZg8dZuoxVDl8A