[PATCH] detect uninitialised per-cpu storage
So poor old Dave spent days hunting down memory corruption because the
`kstat' per-cpu storage is not initialised (it needs to be, it's a workaround
for ancient gcc's).

The same problem had me hunting for a day too.

This patch, based on an initial version from Rusty will
parse System.map at final link and will fail the build if
any per-cpu symbols are found to be not in the percpu section.

BKrev: 3ddd600ddND-0Moe3HWSO2GX4PCF7w