[PATCH] remove a warning from __block_write_full_page()
There is a warning in there to detect when block_write_full_page()
attaches buffers to a blockdev page.  This is a bad thing because that
page's blocks may then overlap blocks from a different address_space.
So I disallowed it.

But the message can be triggered when an application is mmapping a
blockdev MAP_SHARED.  Apparently INND likes to do this.

So remove the warning.

BKrev: 3dde8615n0j0rC4NAnbufYA0ATfEdQ