[PATCH] Allow for profile_buf size = kernel text size
Patch from Ravikiran G Thirumalai <kiran@in.ibm.com>

Allows the kernel profiler to use a shift value of zero for
high-resolution profiling.

So instead of overloading a shift of zero to mean "not profiling", use
a separate flag for that.

This means that anyone who is currently using "profile=0" to disable
profiling has accidentally turned profiling on, so I added a printk.

BKrev: 3de3b5dbQKf3BSqVKPT3Fra94AEQIw